requirements.txt
fastapi
uvicorn
pydantic
cyberwave
python-dotenv